% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_list.R
\name{country_list}
\alias{country_list}
\title{List of countries available in the Covid-19 database}
\usage{
country_list()
}
\description{
This function provides a list of country names inside the Covid-19 database.
The \code{country_name} argument inside the \code{\link{load_covid}} function should be spelled
as they are listed here.
}
\seealso{
\code{\link{load_covid}}
}
