% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pandemicPredicted.R
\name{plot.pandemicPredicted}
\alias{plot.pandemicPredicted}
\title{Plot pandemic predictions}
\usage{
\method{plot}{pandemicPredicted}(object)
}
\arguments{
\item{object}{Output of function \code{\link{posterior_predict.pandemicEstimated}}.}
}
\value{
A list containing two plots, long and short. "long" shows the long-term predictions
and "short" the short-term predictions. If any of them did not get plotted due to lack of prediction,
its value will return NULL.
\item{\code{long}}{
  The plotted long-term prediction. The predictions are made on daily new confirmed cases or daily new deaths.
  }
  \item{\code{short}}{
  The plotted short-term prediction. The predictions are made on daily cumulative cases or daily cumulative deaths.
  }
}
\description{
S3 method that plots the predicted data into an interactive graphic.
}
\examples{
\dontrun{
dataMG = load_covid("Brazil","MG")
estimMG = pandemic_model(dataMG)
predMG = posterior_predict(estimMG)
statsMG = pandemic_stats(predMG)
plot(predMG)}
}
\references{
CovidLP Team, 2020. CovidLP: Short and Long-term Prediction for COVID-19. Departamento de Estatistica. UFMG,
Brazil. URL: \url{http://est.ufmg.br/covidlp/home/en/}
}
\seealso{
\code{\link{posterior_predict.pandemicEstimated}}, \code{\link{pandemic_stats}} and \code{\link{plottedPandemic-objects}}.
}
