% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottedPandemic_Class.R
\name{plottedPandemic-objects}
\alias{plottedPandemic-objects}
\title{plottedPandemic objects: Plots for pandemic data using the plotly library}
\description{
The \pkg{PandemicLP} plotting function returns an object of S3 class
\code{plottedPandemic}, which is a list containing the components described below. \cr
\cr
}
\section{Elements for \code{plottedPandemic} objects}{

\describe{
  \item{\code{long}}{
  The plotted long-term prediction.
  }
  \item{\code{short}}{
  The plotted short-term prediction.
  }
}
}

