% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottedPandemicData_Class.R
\name{plottedPandemicData-objects}
\alias{plottedPandemicData-objects}
\title{plottedPandemicData objects: Plots for pandemic data using the plotly library}
\description{
The \pkg{PandemicLP} plotting function returns an object of S3 class
\code{plottedPandemicData}, which is a list containing the components described below. \cr
\cr
}
\section{Elements for \code{plottedPandemicData} objects}{

\describe{
  \item{\code{new}}{
  The plotted new cases/deaths.
  }
  \item{\code{cumulated.}}{
  The plotted cumulated cases/deaths.
  }
}
}

