% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.pandemicEstimated.R
\name{posterior_predict.pandemicEstimated}
\alias{posterior_predict.pandemicEstimated}
\title{Draw from the posterior predictive distribution for pandemic data}
\usage{
\method{posterior_predict}{pandemicEstimated}(object, horizonLong = 300, horizonShort = 14)
}
\arguments{
\item{object}{An object of class pandemicEstimated created by function stan_pandemic.}

\item{horizonLong}{How far into the future the long term prediction is desired.}

\item{horizonShort}{How far into the future the short term prediction is desired.}
}
\value{
An object of class pandemicPredicted. It includes the sampled predictive distribution,
the model used to predict, which is the same as the one used to estimate the data, and relevant
information for the plot function to be used directly.
}
\description{
The posterior predictive distribution is the distribution of the outcome
implied by the model after using the observed data to update our beliefs
about the unknown parameters in the model. Simulating data from the posterior
predictive distribution using the observed predictors is useful for checking
the fit of the model. Drawing from the posterior predictive distribution at
interesting values of the predictors also lets us visualize how a manipulation
of a predictor affects (a function of) the outcome(s). With new observations of
predictor variables we can use the posterior predictive distribution to generate
predicted outcomes.
}
\seealso{
\code{\link{pandemic_model}} and \code{\link{plot.pandemicPredicted}}.
#' @examples
\dontrun{
dataMG = load_covid("Brazil","MG")
estimMG = pandemic_model(dataMG)
predMG = posterior_predict(estimMG)
predMG}
}
