% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pandemicPredicted.R
\name{print.pandemicPredicted}
\alias{print.pandemicPredicted}
\title{Prints prediction summary of pandemic model}
\usage{
\method{print}{pandemicPredicted}(object, summaryFun = median, printPred = "Long", truncView = 3)
}
\arguments{
\item{summaryFun}{Use this function to summarize the predictions. Default argument is mean, but can be any function on a vector}

\item{printPred}{Valid values are 'Long' or 'Short'. Note that 'Short' will show cumulative predictions}
}
\description{
S3 method designed to summarize the prediction obtained by the
\link{posterior_predict.stanpandemic} function. It is not necessary to call function
directly, only visualize the object directly, unless it is desired to change the default arguments.
}
\examples{
\dontrun{dataMG = load_covid("Brazil","MG")
estimMG = stan_pandemic(dataMG)
predMG = posterior_predict(estimMG)
predMG}
}
\seealso{
\link{posterior_predict.stanpandemic} and \link{plot.pandemicPredicted}
}
