% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pandemicStats.R
\name{print.pandemicStats}
\alias{print.pandemicStats}
\title{Print method for \code{pandemicStats} objects}
\usage{
\method{print}{pandemicStats}(x, ...)
}
\arguments{
\item{x}{An object of S3 class \code{\link{pandemicStats-objects}}.}

\item{...}{Currently unused.}
}
\description{
S3 method designed for \code{pandemicStats} objects. It displays a compact summary of the
95\% credible intervals for the predictions calculated by the pandemic model.
}
\seealso{
\code{\link{pandemic_stats}}
}
