% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.plottedPandemic.R
\name{print.plottedPandemic}
\alias{print.plottedPandemic}
\title{Print method for \code{plottedPandemic} objects}
\usage{
\method{print}{plottedPandemic}(x, ...)
}
\arguments{
\item{x}{An object of S3 class \code{\link{plottedPandemic-objects}}.}

\item{...}{Currently unused.}
}
\value{
Returns \code{x}, invisibly.
}
\description{
The print method for a \code{plottedPandemic} S3 class object displays the single plot or both
separated by a prompt for the user.
}
\details{
The plots generated by \code{\link{plot.pandemicPredicted}} are displayed by this function
If only one of the plots has been generated via the argument \code{term}, then it is displayed
in the Viewer windows as an html object, produced by the \code{\link[plotly]{plot_ly}} function.
If both long term and short term plots are requested, then the long term plot is displayed first
and a message prompts the user to display the next plot, namely the short term one.
}
\seealso{
\code{\link{posterior_predict.pandemicEstimated}} and \code{\link{plot.pandemicPredicted}}
}
