% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.plottedData.R
\name{print.plottedPandemicData}
\alias{print.plottedPandemicData}
\title{Print method for \code{plottedPandemicData} objects}
\usage{
\method{print}{plottedPandemicData}(x, ...)
}
\arguments{
\item{x}{An object of S3 class \code{\link{plottedPandemicData-objects}}.}

\item{...}{Currently unused.}
}
\value{
Returns \code{x}, invisibly.
}
\description{
The print method for a \code{plottedPandemic} S3 class object displays the single plot or both
separated by a prompt for the user.
}
\details{
The plots generated by \code{\link{plot.pandemicData}} are displayed by this function
If only one of the plots has been generated via the argument \code{cases}, then it is displayed
in the Viewer windows as an html object, produced by the \code{\link[plotly]{plot_ly}} function.
If both new and cumulative plots are requested, then the new plot is displayed first
and a message prompts the user to display the next plot, namely the cumulative one.
The plot shows the confirmed cases and the deaths information, but if the user desire to see only one,
he can double click on legend to isolate one trace. Consequently, the axis will be automatically configured
to the scale of the desired trace. To see both again, the user just need to double click again.
}
\seealso{
\code{\link{load_covid}} and \code{\link{plot.pandemicData}}
}
