% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_list.R
\name{state_list}
\alias{state_list}
\title{List of states available in the Covid-19 database}
\usage{
state_list()
}
\description{
This function provides the state abbreviations to be used in the \code{\link{load_covid}} function
and the corresponding state names. Only brazilian states are currently available.
}
\seealso{
\code{\link{load_covid}} and \code{\link{country_list}}.
}
