% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandemic_stats.R
\name{pandemic_stats}
\alias{pandemic_stats}
\title{Relevant Statistics of the Pandemic Model}
\usage{
pandemic_stats(object)
}
\arguments{
\item{object}{an object of S3 class \code{pandemicPredicted} created by function \code{posterior_predict}}
}
\value{
An object of class \code{pandemicStats}. This object is a list containing the following elements:
\itemize{
   \item{\code{data}}{
   A list with a data frame containing the observed pandemic data, a character string with the location name
   and a character string indicating the type of cases predicted.
   }
   \item{\code{ST_predict}}{
   A data frame with the Short-term predictions for the number of cumulative cases. For each future date predicted,
   the mean, median, 2.5 and 97.5 percentiles are provided.
   }
   \item{\code{LT_predict}}{
   A data frame with the Long-term predictions for the number of new cases. For each future date predicted,
   the mean, median, 2.5 and 97.5 percentiles are provided.
   }
   \item{\code{LT_summary}}{
   A list with the estimated total number of cases and the dates for the peak and end of the pandemic.
   In each metric, the median, 2.5 and 97.5 percentiles are provided. For more information, see the
   \strong{Details} section.\cr
   \emph{Warning: These calculations are restricted to the argument \code{horizonLong} in the
   \code{posterior_predict.pandemicEstimated} function. It might be necessary to increase the horizon to in
   order to properly calculate the end of the pandemic and total number of cases.}
   }
   \item{\code{mu}}{
   A data frame with the median values of the mean number of new cases for each date (starting from the first
   observed data point until the last date in the Long-term horizon).
   }
}
}
\description{
This function provides short and long-term predictions for the pandemic. 95% credible intervals are
assigned for every future date predicted, as well as for the total number of cases, and dates for the peak
and end of the pandemic.\cr
Short-term predictions are made on the cumulative counts and are dependent on the previous assignment of the
\code{horizonShort} argument in \code{posterior_predict.pandemicEstimated} function. Long-term predictions
are based on the new cases and the argument \code{horizonLong} in \code{posterior_predict.pandemicEstimated}
function.
}
\details{
\subsection{Total Number of Cases}{
 The total number of cases is obtained by adding the predicted new cases to the cumulative total
 already observed in the data.
 }
 \subsection{Estimated Peak Dates}{
 The 95% credible interval for the peak of cases is selected such that the two limiting dates of the 97.5
 percentile curve coincides with the highest value of the 2.5 percentile curve. This guarantees that all
 possible curves belonging to the confidence band will peak within the defined interval.
 }
 \subsection{End of the Pandemic Dates}{
 Represents the 99 percentile of the total number of cases in the pandemic.
 }
}
\examples{
\dontrun{
italy = load_covid("italy")
estim = pandemic_model(italy, case_type = "confirmed", covidLPconfig = TRUE)
pred = posterior_predict(estim)
stats = pandemic_stats(pred)
stats
}

}
\references{
CovidLPTeam, 2020. CovidLP: Short and Long-term Prediction for COVID-19. Departamento de Estatistica. UFMG,
Brazil. URL: \url{http://est.ufmg.br/covidlp/home/en/}
}
\seealso{
\link{\code{load_covid}} \link{\code{pandemic_model}} \link{\code{posterior_predict.pandemicEstimated}}
\link{\code{plot.pandemicPredicted}} \link{\code{print.pandemicStats}}
}
