% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pandemicPredicted.R
\name{plot.pandemicPredicted}
\alias{plot.pandemicPredicted}
\title{Plot pandemic predictions}
\usage{
plot.pandemicPredicted(object)
}
\arguments{
\item{object}{Output of function posterior_predict.stanpandemic.}
}
\value{
A list containing two plots, long and short. If any of them wasn't plotted due to lack of prediction, its value returns NULL.
}
\description{
S3 method that plots the predicted data into an interactive graphic.
}
\examples{
\dontrun{
dataMG = load_covid("Brazil","MG")
estimMG = stan_pandemic(dataMG)
predMG = posterior_predict(estimMG)
plot(predMG)}
}
\seealso{
\code{\link{posterior_predict.pandemicEstimated}}.
}
