% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pandemicPredicted.R
\name{plot.pandemicPredicted}
\alias{plot.pandemicPredicted}
\title{Plot pandemic predictions}
\usage{
\method{plot}{pandemicPredicted}(object)
}
\arguments{
\item{object}{Output of function posterior_predict.stanpandemic.}
}
\value{
A list containing two plots, long and short. If any of them wasn't plotted due to lack of prediction, its value returns NULL.
\item{\code{long}}{
  The plotted long term prediction. The predictions is made on daily new cases.
  }
  \item{\code{short}}{
  The plotted shirt term prediction. The predictions is made on daily accumulated cases.
  }
}
\description{
S3 method that plots the predicted data into an interactive graphic.
}
\examples{
\dontrun{
dataMG = load_covid("Brazil","MG")
estimMG = stan_pandemic(dataMG)
predMG = posterior_predict(estimMG)
plot(predMG)}
}
\references{
CovidLPTeam, 2020. CovidLP: Short and Long-term Prediction for COVID-19. Departamento de Estatistica. UFMG,
Brazil. URL: \url{http://est.ufmg.br/covidlp/home/en/}
}
\seealso{
\code{\link{posterior_predict.pandemicEstimated}}, \code{\link{pandemic_stats}} and \code{\link{plottedPandemic-objects}}.
}
