% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_pandemicEstimated.R
\name{print.pandemicEstimated}
\alias{print.pandemicEstimated}
\title{Print method for \code{pandemicEstimated} objects}
\usage{
\method{print}{pandemicEstimated}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of S3 class \code{\link[ = pandemicEstimated-objects]{pandemicEstimated}}.}

\item{digits}{Number of digits to use for formatting numbers.}

\item{...}{Currently unused.}
}
\value{
Returns \code{x}, invisibly.
}
\description{
The \code{print} method for \code{pandemicEstimated} object of class S3 displays a compact summary of the
fitted model. See the \strong{Details} section below for descriptions of the different components of the printed
output. \cr
\cr
The \code{fit} component of the \code{\link[ = pandemicEstimated-objects]{pandemicEstimated}} object generated
by the \code{pandemic_model} function can be summarised to provide more information.
Check \code{?`summary,stanfit-method`} for additional summary statistics and diagnostics.
}
\details{
\subsection{Convergence and efficiency diagnostics for Markov Chains}{

Included in the summary are: split effective sample sizes (n_eff), Monte Carlo standard errors
(se_mean) and split Rhats.\cr
\cr
The Monte Carlo standard error provides relevant information for a posterior sample with more than one chain.\cr
\cr
The R-hat convergence diagnostic compares the
between- and within-chain estimates for model parameters and other univariate
quantities of interest. If chains have not mixed well (ie, the between- and
within-chain estimates don't agree), R-hat is larger than 1.
We recommend running at least four chains by default and only using the
sample if R-hat is less than 1.05.

}
\subsection{covidLPconfig}{
This subsection shows the main input settings used by the fitted model, and indicates whether default settings
of the CovidLP app (\url{http://est.ufmg.br/covidlp/home/en/})
were used (\code{covidLPconfig = TRUE} or \code{FALSE}).
Check the default settings of the CovidLP app in \code{\link{pandemic_model}}.
}
\subsection{Priors}{

A list with information about the prior distributions used and model restrictions (if there are any).
In future versions of the package \code{PandemicLP}, the user will be allowed to modify the priors.
For more information, see the \strong{Details} section of \code{\link{pandemic_model}}.


}
}
\seealso{
\code{?`summary,stanfit-method`}
}
